HtmlAreaContext is missing in some render locations #10406
Centralized `HtmlAreaProvider` mount in `FormRenderer` via `HtmlAreaShell` to ensure context is available wherever forms are rendered.
Added `useOptionalHtmlAreaContext` to skip nested provider mounts when a parent already supplies the context.
Moved `useApplicationKeys` from `content-wizard-tabs/` to shared `features/hooks/` and updated import sites.
Removed redundant `HtmlAreaProvider` wrappers from `ContentForm`, `MixinView`, and `ComponentInspectionPanel`.
