+++ modules/lib/src/main/resources/assets/js/v6/features/hooks/useApplicationKeys.ts
@@ -1,10 +1,10 @@
 import type {ApplicationKey} from '@enonic/lib-admin-ui/application/ApplicationKey';
 import {useStore} from '@nanostores/preact';
 import {useEffect, useMemo, useState} from 'react';
-import type {Site} from '../../../../../app/content/Site';
-import {loadNearestSite} from '../../../api/details';
-import {$activeProject} from '../../../store/projects.store';
-import {$contextContent} from '../../../store/context/contextContent.store';
+import type {Site} from '../../../app/content/Site';
+import {loadNearestSite} from '../api/details';
+import {$activeProject} from '../store/projects.store';
+import {$contextContent} from '../store/context/contextContent.store';
 
 export function useApplicationKeys(): ApplicationKey[] {
     const contextContent = useStore($contextContent);
