+++ modules/lib/src/main/resources/assets/js/v6/features/shared/form/input-types/html-area/HtmlAreaContext.tsx
@@ -46,3 +46,7 @@ export const useHtmlAreaContext = (): HtmlAreaContextValue => {
 
     return context;
 };
+
+export const useOptionalHtmlAreaContext = (): HtmlAreaContextValue | undefined => {
+    return useContext(HtmlAreaContext);
+};
