+++ modules/lib/src/main/resources/assets/js/v6/features/shared/form/input-types/html-area/index.ts
@@ -1,4 +1,4 @@
 export type {HtmlAreaConfig} from './HtmlAreaConfig';
 export {HtmlAreaDescriptor} from './HtmlAreaDescriptor';
 export {HtmlAreaInput} from './HtmlAreaInput';
-export {HtmlAreaProvider, useHtmlAreaContext} from './HtmlAreaContext';
+export {HtmlAreaProvider, useHtmlAreaContext, useOptionalHtmlAreaContext} from './HtmlAreaContext';
