+++ modules/lib/src/main/resources/assets/js/v6/features/views/wizard/content-wizard-tabs/ContentForm.tsx
@@ -1,25 +1,17 @@
 import {RawValueProvider, ValidationVisibilityProvider} from '@enonic/lib-admin-ui/form2';
-import {CONFIG} from '@enonic/lib-admin-ui/util/Config';
 import {useStore} from '@nanostores/preact';
 import {type ReactElement, useEffect, useMemo} from 'react';
 import {FormRenderer} from '../../../shared/form';
-import {HtmlAreaProvider} from '../../../shared/form/input-types/html-area';
-import {$contextContent} from '../../../store/context/contextContent.store';
-import {$activeProject} from '../../../store/projects.store';
 import {$contentType, $wizardDraftData, notifyContentFormMounted} from '../../../store/wizardContent.store';
 import {$validationVisibility, getContentRawValueMap} from '../../../store/wizardValidation.store';
 import {DisplayNameInput} from './DisplayNameInput';
-import {useApplicationKeys} from './useApplicationKeys';
 
 const CONTENT_FORM_NAME = 'ContentForm';
 
 export const ContentForm = (): ReactElement | null => {
     const contentType = useStore($contentType);
     const draftData = useStore($wizardDraftData);
-    const contextContent = useStore($contextContent);
-    const activeProject = useStore($activeProject);
     const visibility = useStore($validationVisibility);
-    const applicationKeys = useApplicationKeys();
 
     const isReady = contentType != null && draftData != null;
 
@@ -31,14 +23,10 @@ export const ContentForm = (): ReactElement | null => {
 
     const rawValueMap = useMemo(() => getContentRawValueMap(), []);
 
-    const contentSummary = contextContent ?? undefined;
-
     const applicationKey = useMemo(
         () => contentType?.getContentTypeName().getApplicationKey(),
         [contentType],
     );
-
-    const assetsUri = CONFIG.getString('assetsUri');
 
     if (!isReady) {
         return null;
@@ -49,18 +37,11 @@ export const ContentForm = (): ReactElement | null => {
             <DisplayNameInput />
             <ValidationVisibilityProvider visibility={visibility}>
                 <RawValueProvider map={rawValueMap}>
-                    <HtmlAreaProvider
-                        contentSummary={contentSummary}
-                        project={activeProject}
-                        applicationKeys={applicationKeys}
-                        assetsUri={assetsUri}
-                    >
-                        <FormRenderer
-                            form={contentType.getForm()}
-                            propertySet={draftData.getRoot()}
-                            applicationKey={applicationKey}
-                        />
-                    </HtmlAreaProvider>
+                    <FormRenderer
+                        form={contentType.getForm()}
+                        propertySet={draftData.getRoot()}
+                        applicationKey={applicationKey}
+                    />
                 </RawValueProvider>
             </ValidationVisibilityProvider>
         </div>
